Part 1 of MSVC /W3 conformancy. Signed/Unsigned mismatch fixes and type conversion warnings.

+++ src/p_saveg.c
@@ -558,7 +558,7 @@ static void saveg_read_ticcmd_t(ticcmd_t *str)
     str->angleturn = saveg_read16();
 
     // short consistancy;
-    str->consistancy = saveg_read16();
+    str->consistancy = (byte)saveg_read16();
 
     // byte chatchar;
     str->chatchar = saveg_read8();
